requirements.txt
aiohttp==3.8.1
nest-asyncio==1.5.4
asyncio==3.4.3
newspaper3k==0.2.8
sklearn
numpy
matplotlib
openai
voila
ipywidgets